Add modifying assignment.
Possibly unclear:

  a b: 1 10
  -> a -> b + 1

What's the value of b?

If we parse as ->a->b +1, then b is 2 (a->b is an as-pattern).
If we parse as ->a ->b+1, then b is 1 (a is the pattern, ->b is a forward assignment).

Currently we parse as the first one, which is maybe not so good because I don't think there's a way to disambiguate to the second one if you want to.