Add branching.
ALl values are considered truthy unless the first atom they contain is 0
(can be char 0, integer 0, or floating 0, but maybe char 0 is a bit much).

`if` expressions come in two flavors:
  - Two-argument `If`, which acts as a predicate taking a single boolean left argument.
      e.g. Fac: { x = 0 If(1; x * (x - 1 Rec)) }
  - Three-argument `if`, which is parsed like a small noun and takes the condition as the first argument.
      e.g. Fac: { if(x = 0; 1; x * (x - 1 Rec)) }
    Like other small nouns, these can be stranded:
      e.g. (x:5) if(x>100; 100; x) 17
      gives 5 5 17

Before, we would have had to write factorial something like this:
  Fac: {
    fac: rec
    x = 0 ~@ [{x * (x - 1 Fac)}; .1] @ x
  }
Yuck!

We also eliminate tail calls made from within arbitrarily-nested ifs.

Currently, locals defined within the then- or else-branch of an `if` are not
visible outside of it, but maybe we can make a variable visible if it's defined
in all branches. All instances would need to occupy the same slot, or we move
them to the same one afterwards.