Syntax changes! Now, parentheses don't always create a noun; in a block (e0; e1; ...; eN), the block's overall part of speech is now that of eN. This makes trains less burdensome, since we can parenthesize verbs to associate them properly without littering code with dots. Trains are supposed to be convenient.
With that change alone, it would no longer be easy to turn a lambda into a noun;
before, you could use `({...})`, but now, you'd have to assign it and use a
lowercase name as in `F: {...}; f` or put it in another nounifying context as in
`[{...}]@0` or `[]{{...}}`. The former can litter the namespace and the latter
are too obtuse, so I've added an adverb-like syntax: `_(e)' is the same as
`(e)`, simply parsed as a noun. So `f: _{x+1}` is valid, and `5 {x Y} _{x*2}`
results in 10. Note that you can also strand these, so `1 _{x%2} 3 _>: _`:#` is
an array of 5 elements (though I'd write the same array as [1; {x%2}; 3; >:;
`:#]).

The underscore can be thought of as producing a constant function in trains,
since using a noun in a train binds it as the right argument of a verb. `3 (P {x
Y Y} _{x*3})` = `3 ({x Y Y} _{x*3})` = 27.

The parsing algorithm for blocks is pretty bad now. When we see a (, we don't
know if the whole block will be a noun or a verb until we parse its last
expression, and currently we just parse the whole thing and backtrack if it
isn't what we're looking for. I have some TODOs to improve it.