Allow user-defined functions to be used as adverbs.
Syntax:
  x F[v1; v2; ...; vN] y
In this context, F is treated as an adverb with N operands, invoked as a dyadic verb on x and y. Currently, F must be defined to accept N arguments and return a
function. F can also be a lambda or a (parenthesized) verb block.

Example REPL session:
    Fold: {|v| {if(x#=0; y; (next: y V (x@0); 1-(x#)~#x Rec next))}}
    1 2 3 Fold[+] 0
  6

The indexing syntax also applies to nouns:
  a[i1; i2; ...; iN]
This denotes a progressive indexing; a is first indexed by each atom in i1, and then the result of each atom index is further indexed by i2, and so on.

Example REPL session:
    a:[1 2 3
  ↪    4 5
  ↪    [6 7; 8]]
    a[0 2; 1]
  2 8
    a[0 2; 0]
  [1
   6 7]

In general, an indexed expression e[...] inherits e's part of speech.

Caveats and gotchas:
  - Indexing is differentiated from array literal syntax by directly following
  an expression, with no whitespace in-between. In "x F[a]", [a] indexes F, and
  in "x F [a]", [a] is an array literal.
  - Indexing binds more tightly than stranding, so 1 2 3[0] is 1 2 3, and 1 2
  3[1] is an index error.
  - Indexing doesn't currently apply to primitive verbs or adverbs, so "x p[a]",
  where p is a primitive, is the same as "x p [a]".