Support multiple cases in explicit functions. Syntax:     F: {|x| "I got one arg!" : |x;y| "I got two!" : |x;y;z| "Wow, a whole three of them!"}     1 F   "I got one arg!"     f[1]   "I got one arg!"     1 F 2   "I got two!"     f[1;2]   "I got two!"     f[1;2;3]   "Wow, a whole three of them!"
This can be used as a pattern matching mechanism:
    1 2 3 {
    : |[a]| "One"
    : |a b c| "Three"
    : |a b| "Two"
    }
  "Three"

A colon before the first case is allowed but unnecessary.

There's still a whole lot to clean up around function arguments and
calling; call_val and index_or_call should be merged, and I also don't think
having saved_args in StackFrame is the smoothest thing.

I've also removed AmbivalentCases, since supporting it in the presence of
argument lists of arbitrary length didn't seem worth it; this feature can
supersede it.