Fix progressive_index.
Before, when arguments were shorter than their predecessors, progressive_index
would panic or reach too deep into the stack, because we were popping for each
element of the preceding index.

      ((2 3 4) (4 5 6))[0 1; 2]
    thread 'main' panicked at src/vm.rs:518:26:
    called `Option::unwrap()` on a `None` value

Here, we indexed by 0, then popped 2 to index further; when we back up to
process index 1, the 2 is no longer present, leading to a failed pop.

Now, instead of popping, we just just index the stack and truncate it when
progressive_index is done.