Support partially-applied function cases in the VM.
This introduces a new type, ArgSpec, that consists of an arity and a mask of the
arguments expected by / provided to a function.

Also, two new instructions: CallSpec for calling and ArgCheck for Eventually, I
think we'll replace some or all of the existing Call instructions with CallSpec
and ArgCheckEq with ArgCheck.