Fix a bug in parsing view patterns.
Before, SmallVerb -> Pattern was unintentionally considered "small", letting it
strand with other patterns; for example, `x F -> a G -> b` was parsed
equivalently to `x F -> [a; G -> b]` when it should be equivalent to
`(a: x F; a G -> b)`.

Concrete example:
  "babcab" = "b" ->eq \:+ ->acc
Before:
  Array unpacking failed; expected 2 elements, got 6
After:
  1 1 2 2 2 3

Also, we now backtrack to allow view patterns of the form (verb)->pattern.