Fix the arity of p/q's operand.
The operand of p or q should both have arity 1 regardless of the derived verb's arity.
    IsDigit: {x (~? =! q#) "0123456789"}
    isDigit["8a2390s0slls9"]
Before:
  thread 'main' panicked at src/vm.rs:1062:18:
  not yet implemented: U8s([48, 49, 50, 51, 52, 53, 54, 55, 56, 57]) Take
After:
  0

Also, revert print debugging from parsing; originally I was debugging the
following parse error:
    F: {[ [ l1; ..a -> c ] ] l1}
  Unexpected `->'; expected `]'
But actually subarray patterns (..a) aren't allowed as the lhs of an as-pattern
anyway. This is intended, actually...! Obviously needs a better error message